+++ packages/lu/test/parser/luis/luisBuilder.test.js
@@ -58,4 +58,46 @@ assert.isTrue(luisObject.validate())
         assert.equal(luisObject.entities[0].roles.length, 2);
         assert.deepEqual(luisObject.entities[0].roles, ['firstName', 'lastName']);
     });
+
+    it('Parse to LU instance', async () => {
+        let luFile = `
+        @ ml test
+        # test
+        - this is a {@test = one}
+        `;
+
+        let result = `
+> LUIS application information
+> !# @app.versionId = 0.1
+> !# @app.culture = en-us
+> !# @app.luis_schema_version = 3.2.0
+
+
+> # Intent definitions
+
+## test
+- this is a {@test=one}
+
+
+> # Entity definitions
+
+@ ml test
+
+
+> # PREBUILT Entity definitions
+
+
+> # Phrase list definitions
+
+
+> # List entities
+
+> # RegEx entities
+
+
+`
+        const luisObject = await LUISBuilder.fromLUAsync(luFile)
+        const newLU = luisObject.parseToLU()
+        assert.equal(newLU.content, result); 
+    });
 });
